Factor out Sampler type
