+++ src/lib.rs
@@ -35,6 +35,7 @@ use anyhow::{Error, Result};
 
 mod core;
 pub mod recorder;
+mod sampler;
 mod storage;
 mod ui;
