+++ src/main.rs
@@ -17,6 +17,7 @@ use std::fs::DirBuilder;
 use std::os::unix::prelude::*;
 use std::path::{Path, PathBuf};
 use std::process::Command;
+use std::sync::atomic::{AtomicBool, Ordering};
 
 /// The kinds of things we can call `rbspy record` on.
 #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
@@ -159,6 +160,8 @@ fn do_main() -> Result<(), Error> {
                 }
             };
 
+            let done = std::sync::Arc::<AtomicBool>::new(AtomicBool::new(false));
+            let done_handler = done.clone();
             let config = recorder::RecordConfig {
                 format,
                 raw_path: raw_path.clone(),
@@ -170,7 +173,21 @@ fn do_main() -> Result<(), Error> {
                 maybe_duration,
                 flame_min_width,
                 lock_process,
+                done,
             };
+
+            ctrlc::set_handler(move || {
+                if done_handler.load(Ordering::Relaxed) {
+                    eprintln!("Multiple interrupts received, exiting with haste!");
+                    std::process::exit(1);
+                }
+                eprintln!("Interrupted.");
+                // Trigger the end of the loop
+                done_handler.store(true, Ordering::Relaxed);
+            })
+            .expect("Error setting Ctrl-C handler");
+
+            eprintln!("Press Ctrl+C to stop");
 
             let output_paths_message = format!(
                 "Wrote raw data to {}\nWrote formatted output to {}",
@@ -405,6 +422,7 @@ impl Args {
         Args::from(env::args())
     }
 }
+
 fn output_filename(
     base_dir: &str,
     maybe_filename: Option<&str>,
