+++ src/recorder/mod.rs
@@ -1,6 +1,6 @@
 mod record;
 mod snapshot;
 
-pub use record::parallel_record as record;
 pub use record::Config as RecordConfig;
+pub use record::record;
 pub use snapshot::snapshot;
